requirements.txt
fastapi
pydantic
python-multipart
uvicorn
openai
sentence-transformers
faiss-cpu
pytesseract
Pillow
python-dotenv
